+++ server/controllers/surveys.js
@@ -9,7 +9,18 @@ module.exports = {
       if(err) {
         console.log('Error reading survey definitions:', err);
       }
-      const surveys = _.map(files, (filePath) => {
+      if(process.env.SURVEY) {
+        files = _.filter(files, (file) => {
+          
+          // Check if the name matches with or without extension
+          const fullName = file.replace('surveys/', '');
+          const name = fullName.replace('.json', '');
+          
+          return fullName === process.env.SURVEY || name === process.env.SURVEY;
+        });
+      }      
+
+      let surveys = _.map(files, (filePath) => {
         const survey = require(path.join('..', filePath));
         survey.id = idFromString(survey.title);
         return survey;