+++ changelog.md
@@ -212,6 +212,7 @@ More data saving and map activity features.
 * Add architecture docs
 * Make home redirect to the first survey
 * Add a survey name variable to docker build
+* Make `/surveys` endpoint only return the specified survey when specified
 
 I used a couple blog posts about how to set up Express and postgres:
 * https://medium.freecodecamp.org/how-to-make-create-react-app-work-with-a-node-backend-api-7c5c48acb1b0